fix: remove not needed pkg
