+++ internal/provisioner/ansible.go
@@ -11,7 +11,6 @@ import (
 
 	"github.com/celestiaorg/talis/internal/events"
 	"github.com/celestiaorg/talis/internal/logger"
-	"github.com/celestiaorg/talis/internal/provisioner/config"
 	"github.com/celestiaorg/talis/internal/types"
 )
 
@@ -39,14 +38,51 @@ const (
 
 var _ Provisioner = (*AnsibleProvisioner)(nil)
 
+// AnsibleConfig represents the configuration for Ansible provisioner
+type AnsibleConfig struct {
+	// JobID is the unique identifier for the current job
+	JobID string
+	// JobName is the name of the current job
+	JobName string
+	// OwnerID is the ID of the owner of the instances
+	OwnerID uint
+	// SSHKeyPath is the path to the SSH private key
+	SSHKeyPath string
+	// SSHUser is the user for SSH connections
+	SSHUser string
+	// PlaybookPath is the path to the Ansible playbook
+	PlaybookPath string
+	// InventoryBasePath is the base path for inventory files
+	InventoryBasePath string
+}
+
+// Validate validates the Ansible configuration
+func (c *AnsibleConfig) Validate() error {
+	// Validate required fields
+	if c.JobID == "" {
+		return fmt.Errorf("JobID is required")
+	}
+	// TODO: Add validation for OwnerID
+	// if c.OwnerID == 0 {
+	// 	return fmt.Errorf("OwnerID is required")
+	// }
+
+	// Set default values for optional fields
+	if c.InventoryBasePath == "" {
+		c.InventoryBasePath = InventoryBasePath
+	}
+
+	return nil
+}
+
 // AnsibleProvisioner implements provisioning using Ansible
 type AnsibleProvisioner struct {
-	config     *config.AnsibleConfig
+	config     *AnsibleConfig
 	sshChecker SSHChecker
 }
 
 // NewAnsibleProvisioner creates a new Ansible provisioner
-func NewAnsibleProvisioner(config *config.AnsibleConfig) (*AnsibleProvisioner, error) {
+func NewAnsibleProvisioner(config *AnsibleConfig) (*AnsibleProvisioner, error) {
 	if err := config.Validate(); err != nil {
 		return nil, fmt.Errorf("invalid config: %w", err)
 	}
